Remove size parameter for shape and text constructors
Size is redundant for shapes since they all have parameters to set the size, leading to confusion. This hard codes all sizes to 100% for shapes.

Also would cause issues when attempting to use physics with shapes that have the size parameter set, as for shapes size is ignored in the physics engine